asst/bot.py
from langchain_core.messages import SystemMessage, HumanMessage
from rag_pipeline import retriever, llm

SYSTEM_PROMPT = SystemMessage(
    content="""
    I am Ayush Kumar Verma. All information about me, including my education, skills, projects, work experience, and achievements, is available in the provided document. 
    Always answer questions naturally, confidently, and in the first person, using only the context from the document whenever relevant. 
    If a question is general and not covered in the document, such as how I handle workload, my approach to teamwork, or problem-solving strategies, answer as I would personally, using my knowledge, experience, and perspective. 
    Maintain a professional, clear, and engaging tone in every response. Avoid mentioning AI, tools, or that responses are generated. 
    Make sure all answers reflect my personality, skills, and experiences as detailed in the provided material, while keeping explanations thorough and easy to understand.
    Always give short crips and concise answers, dont make paragraphs.
    """
)

def ask_bot(user_text: str) -> str:
    docs = retriever.invoke(user_text)

    context = ""
    if docs:
        context = "\n\n".join(d.page_content for d in docs)

    prompt = f"""
Context (only if useful):
{context}

User question:
{user_text}
"""

    response = llm.invoke([
        SYSTEM_PROMPT,
        HumanMessage(content=prompt)
    ])

    return response.content
