requirements.txt
streamlit
streamlit-mic-recorder
groq
langchain
langgraph
faiss-cpu
gtts
python-dotenv
pypdf
